Remove PwshTaskBuilder
