+++ src/Sharpliner/AzureDevOps/AzureDevOpsPipelineDefinition.cs
@@ -25,8 +25,8 @@ namespace Sharpliner.AzureDevOps
         protected static ConditionedDefinition<VariableBase> Variable(string name, bool value) => new(new Variable(name, value));
         protected static ConditionedDefinition<VariableBase> Group(string name) => new(new VariableGroup(name));
         protected static BashTaskBuilder Bash { get; } = new();
-        protected static PowershellTaskBuilder Powershell { get; } = new();
-        protected static PwshTaskBuilder Pwsh { get; } = new();
+        protected static PowershellTaskBuilder Powershell { get; } = new(false);
+        protected static PowershellTaskBuilder Pwsh { get; } = new(true);
         protected static PublishTask Publish(string displayName, string filePath) => new(displayName, filePath);
         protected static PublishTask Publish(string filePath) => new(null, filePath);
         protected static CheckoutTaskBuilder Checkout { get; } = new();
