+++ src/Sharpliner/AzureDevOps/Tasks/Builders/PowerShellTaskBuilder.cs
@@ -4,6 +4,8 @@ namespace Sharpliner.AzureDevOps.Tasks
 {
     public class PowershellTaskBuilder : ScriptTaskBuilder
     {
+        private readonly bool _pwsh;
+
         /// <summary>
         /// Creates a Powershell task where the contents come from an embedded resource.
         /// </summary>
@@ -11,7 +13,10 @@ namespace Sharpliner.AzureDevOps.Tasks
         /// <param name="displayName">Name of the build step</param>
         /// <param name="resourceFileName">Name of the resource file</param>
         public InlinePowershellTask FromResourceFile(string displayName, string resourceFileName)
-            => new(displayName, GetResourceFile(Assembly.GetCallingAssembly()!, resourceFileName));
+            => new InlinePowershellTask(displayName, GetResourceFile(Assembly.GetCallingAssembly()!, resourceFileName)) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task where the contents come from an embedded resource.
@@ -19,7 +24,10 @@ namespace Sharpliner.AzureDevOps.Tasks
         /// <typeparam name="TAssembly">A type located in the assembly where the resource is located</typeparam>
         /// <param name="resourceFileName">Name of the resource file</param>
         public InlinePowershellTask FromResourceFile(string resourceFileName)
-            => new(null, GetResourceFile(Assembly.GetCallingAssembly()!, resourceFileName));
+            => new InlinePowershellTask(null, GetResourceFile(Assembly.GetCallingAssembly()!, resourceFileName)) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task where the contents come from a file.
@@ -28,43 +36,67 @@ namespace Sharpliner.AzureDevOps.Tasks
         /// <param name="displayName">Name of the build step</param>
         /// <param name="path">Path to the file</param>
         public InlinePowershellTask FromFile(string displayName, string path)
-            => new(displayName, System.IO.File.ReadAllText(path));
+            => new InlinePowershellTask(displayName, System.IO.File.ReadAllText(path)) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task where the contents come from a file.
         /// The contents are inlined in the YAML as contrary to File method where the file name is just referenced.
         /// </summary>
         /// <param name="path">Path to the file</param>
-        public InlinePowershellTask FromFile(string path) => new(null, System.IO.File.ReadAllText(path));
+        public InlinePowershellTask FromFile(string path)
+            => new InlinePowershellTask(null, System.IO.File.ReadAllText(path)) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task referencing a Powershell file (contents are not inlined in the YAML).
         /// </summary>
         /// <param name="displayName">Name of the build step</param>
         /// <param name="filePath">Path to the file</param>
-        public PowershellFileTask File(string displayName, string filePath) => new(displayName, filePath);
+        public PowershellFileTask File(string displayName, string filePath)
+            => new PowershellFileTask(displayName, filePath) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task referencing a Powershell file (contents are not inlined in the YAML).
         /// </summary>
         /// <param name="filePath">Path to the file</param>
-        public PowershellFileTask File(string filePath) => new(null!, filePath);
+        public PowershellFileTask File(string filePath)
+            => new PowershellFileTask(null!, filePath) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task with given contents.
         /// </summary>
         /// <param name="displayName">Name of the build step</param>
         /// <param name="scriptLines">Contents of the script</param>
-        public InlinePowershellTask Inline(string displayName, params string[] scriptLines) => new(displayName, scriptLines);
+        public InlinePowershellTask Inline(string displayName, params string[] scriptLines)
+            => new InlinePowershellTask(displayName, scriptLines) with
+            {
+                Pwsh = _pwsh
+            };
 
         /// <summary>
         /// Creates a Powershell task with given contents.
         /// </summary>
         /// <param name="scriptLines">Contents of the script</param>
-        public InlinePowershellTask Inline(params string[] scriptLines) => new(null!, scriptLines);
+        public InlinePowershellTask Inline(params string[] scriptLines)
+            => new InlinePowershellTask(null!, scriptLines) with
+            {
+                Pwsh = _pwsh
+            };
 
-        internal PowershellTaskBuilder()
+        internal PowershellTaskBuilder(bool pwsh)
         {
+            _pwsh = pwsh;
         }
     }
 }
